requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
beautifulsoup4==4.11.1
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==3.0.1
idna==3.4; python_version >= '3.5'
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1; python_version >= '3.6'
urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
